Remove remaining `WriteRepository`
- Refactor `WriteRepository` into independent `DiscoveryEngine::Put` and
  `DiscoveryEngine::Delete` services
- Update `PublishingApiDocument` classes to call new services instead of
  repository
- Add `DiscoveryEngine::DocumentName` helper module to handle Discovery
  Engine document (path) names
- Add Discovery Engine datastore path to application configuration
- Move and update integration test to reflect new repository-less
  testing philosophy
