+++ app/models/publishing_api_document/unpublish.rb
@@ -1,8 +1,8 @@
 module PublishingApiDocument
   class Unpublish < Base
-    # Synchronize the document to the given repository (i.e. delete it from the repository).
-    def synchronize_to(repository)
-      repository.delete(content_id, payload_version:)
+    # Synchronize the document to the given service (i.e. delete it remotely).
+    def synchronize(service: DiscoveryEngine::Delete.new)
+      service.call(content_id, payload_version:)
     end
   end
 end
