+++ config/environments/test.rb
@@ -1,5 +1,4 @@
 require "active_support/core_ext/integer/time"
-require "repositories/test_repository"
 
 # The test environment is used exclusively to run your application's
 # test suite. You never need to work with it otherwise. Remember that
